Implemented sharing receiver activity
The receiver activity parses the share request, and if it is valid
forwards it on to the Edit Reference activity.

Refactored Edit Reference to support receiving arbitrary default values,
and to support selecting a Project if no project Id has been provided.
(This is currently only available for new References)