Updated to latest ADAL (1.0.3)
Login screen now has a proper splash theme, to prevent the screen flash on
application start

Refactored AuthManager to make proper use of the ADAL's auth token
caching. This is in place of our own hand-rolled token caching code which
made no attempt at encryption. (note: this is faulty in the emulator which
seems to have issues with encryption).

No longer passing a login hint.
(Seems to be used in cache key generation? Not what we want.)

If the user taps "sign in" they are now always prompted for their
password. If the user has signed in previously we will attempt to
auto-sign in when the application launches.